create empty state.json if no docker socket found

+++ entrypoint.sh
@@ -8,7 +8,11 @@ fi
 
 if [ ! -f /App/Frierun/state.json ]; then
   echo "No state.json file found, creating one"
-  cp /App/state.json.default /App/Frierun/state.json
+  if [ -f /var/run/docker.sock ]; then 
+    cp /App/state.json.default /App/Frierun/state.json
+  else
+    touch /App/Frierun/state.json
+  fi
 fi
 
 exec "$@"
